refactor: extract Plone template family to @volto-hydra/helpers
Move 21 exports from packages/hydra-js/hydra.src.js to
packages/helpers/index.js. All are pure data helpers — no DOM,
no postMessage, no Bridge state — so they belong with the rest of
the SSR-safe template/listing toolkit. No re-exports added in
hydra.src.js — direct caller updates only.

Moved exports:
  - TEMPLATE_MARKER
  - isLayoutTemplate, findSlotRegions, isTemplateAllowedIn
  - getLayoutTemplates, getSnippetTemplates, getTemplateBlocks
  - isFixedTemplateBlock, isPlaceholderContent
  - templateIdToPath, getUniqueTemplateIds
  - isBlockInEditedTemplate, isBlockReadonly, isBlockPositionLocked
  - getBlockAddability, canContainAll
  - cloneBlocksWithNewIds, insertSnippetBlocks
  - loadTemplates, expandTemplates, expandTemplatesSync

Stays in hydra-js (browser-coupled or admin-only):
  - Bridge class, initBridge
  - getAccessToken, getTokenFromCookie, getAuthHeaders, isEditMode
  - calculateDragHandlePosition (admin UI positioning)
  - PAGE_BLOCK_UID (too widely used; follow-up)
  - canContain, findConversionPath, mapLayoutItems (containerOps.js)

Adaptations beyond mechanical move:
  - canContainAll's body moved out of containerOps.js (only used by
    iframe + admin DnD); other containerOps stay put. containerOps.test.js
    now imports canContainAll from @volto-hydra/helpers.
  - expandTemplatesSync used hydra-js's isEditMode; the helpers copy
    uses a private _isEditMode() mirroring the same logic SSR-safely.
  - hydra.src.js imports the predicates it still calls internally
    (isBlockReadonly, isBlockPositionLocked, getBlockAddability) from
    @volto-hydra/helpers; no user-facing re-exports.

Files updated (20):
  - packages/helpers/index.js: +1490 lines (template family)
  - packages/hydra-js/hydra.src.js: -1490 lines (template family removed,
    internal imports added)
  - packages/hydra-js/containerOps.js: -22 lines (canContainAll removed)
  - packages/hydra-js/{applyLayoutTemplate,containerOps,
    expandTemplatesObjectList}.test.js: import from @volto-hydra/helpers
  - packages/volto-hydra/src/utils/{blockPath.js, mergeTemplates.mjs},
    packages/volto-hydra/src/components/{Iframe/View.jsx,
    Sidebar/{ChildBlocksWidget,LayoutSelector,ParentBlocksWidget}.jsx,
    Toolbar/SyncedSlateToolbar.jsx}: split imports
  - examples/nuxt-blog-starter/{components/{AccordionBlock,block}.vue,
    composables/ploneApi.js, pages/[...slug].vue}: split imports
  - examples/hydra-vue-f7/src/{js/routes.js, components/block.vue}: split
  - docs/examples/test-react/src/main.jsx: split

Bundle size:
  packages/hydra-js/hydra.js: 439kb → 408.1kb (-30.9kb)

Test results:
  - hydra-js jest: 54/54 pass
  - doc-examples playwright (astro, svelte, react, vue): 80/80 pass
