+++ examples/hydra-vue-f7/src/js/routes.js
@@ -2,7 +2,8 @@
 import Page from '../pages/page.vue';
 import NotFoundPage from '../pages/404.vue';
 import axios from 'axios';
-import { loadTemplates, isEditMode } from '@hydra-js/hydra.js';
+import { isEditMode } from '@hydra-js/hydra.js';
+import { loadTemplates } from '@hydra-js/helpers';
 
 // Shared template cache across route navigations
 const templateCache = {};
