+++ packages/volto-hydra/src/components/Toolbar/SyncedSlateToolbar.jsx
@@ -9,8 +9,8 @@ import { BlockButton } from '@plone/volto-slate/editor/ui';
 import slateTransforms, { withEmptyInlineRemoval } from '../../utils/slateTransforms';
 import { syncCreateSlateBlock } from '@plone/volto-slate/utils/volto-blocks';
 import { getBlockById, updateBlockById } from '../../utils/blockPath';
-import { calculateDragHandlePosition, PAGE_BLOCK_UID, isBlockPositionLocked, isBlockReadonly } from '@volto-hydra/hydra-js';
-import { isSlateFieldType } from '@volto-hydra/helpers';
+import { calculateDragHandlePosition, PAGE_BLOCK_UID } from '@volto-hydra/hydra-js';
+import { isSlateFieldType, isBlockPositionLocked, isBlockReadonly } from '@volto-hydra/helpers';
 import { useDispatch, useSelector } from 'react-redux';
 import FormatDropdown from './FormatDropdown';
 import DropdownMenu from './DropdownMenu';
